Updates from UX
Implements changes from UX testing.  All the dynamic components are now clustered in the righthand corner, with one column for the “step value of change” (rather than sensitivity) and the next column for the target and current amounts.  These values revert back to a fixed position and the up and down arrows will change to green depending on which direction is recommended.  On completion on an activity both up and down arrows turn green and does the activity indicator giving another cue to select the next activity.